# Use std::unordered_map::emplace instead of operator[] when producing JSONs