Micro performance improvement...




<!-- BENCHMARK_RESULTS_START -->
<details><summary><h2>Benchmark Results</h2></summary>

| Benchmark | Base | PR |
|-----------|------|----|
| alias | aliased u32: 1124.68<br/>plain u32: 1134.26<br/>aliased double: 970.269<br/>plain double: 981.15 | aliased u32: 1089.78<br/>plain u32: 1091.51<br/>aliased double: 989.96<br/>plain double: 960.508 |
| e2e_match_ch | Total: 2892.8239345550537ms<br/>Min time: 2.416849136352539ms<br/>Mean time: 22.082625454618732ms<br/>Median time: 18.298625946044922ms<br/>95th percentile: 68.8619613647461ms<br/>99th percentile: 82.38818645477294ms<br/>Max time: 93.06883811950684ms | Total: 2908.677101135254ms<br/>Min time: 2.4111270904541016ms<br/>Mean time: 22.203641993398886ms<br/>Median time: 16.501903533935547ms<br/>95th percentile: 70.01745700836182ms<br/>99th percentile: 84.45518016815183ms<br/>Max time: 97.4118709564209ms |
| e2e_match_mld | Total: 2049.659252166748ms<br/>Min time: 1.9621849060058594ms<br/>Mean time: 15.646253833333954ms<br/>Median time: 9.110689163208008ms<br/>95th percentile: 50.00472068786621ms<br/>99th percentile: 59.01558399200434ms<br/>Max time: 67.37112998962402ms | Total: 2021.8911170959473ms<br/>Min time: 2.1300315856933594ms<br/>Mean time: 15.434283336610285ms<br/>Median time: 8.893489837646484ms<br/>95th percentile: 49.703240394592285ms<br/>99th percentile: 57.58173465728754ms<br/>Max time: 67.54612922668457ms |
| e2e_nearest_ch | Total: 1363.0824089050293ms<br/>Min time: 1.1556148529052734ms<br/>Mean time: 1.3630824089050293ms<br/>Median time: 1.2770891189575195ms<br/>95th percentile: 1.7802834510803223ms<br/>99th percentile: 1.8301677703857422ms<br/>Max time: 1.9445419311523438ms | Total: 1371.1118698120117ms<br/>Min time: 1.1625289916992188ms<br/>Mean time: 1.3711118698120117ms<br/>Median time: 1.2840032577514648ms<br/>95th percentile: 1.7879009246826172ms<br/>99th percentile: 1.8647265434265137ms<br/>Max time: 1.9659996032714844ms |
| e2e_nearest_mld | Total: 1367.5897121429443ms<br/>Min time: 1.1601448059082031ms<br/>Mean time: 1.3675897121429443ms<br/>Median time: 1.2780427932739258ms<br/>95th percentile: 1.7817378044128418ms<br/>99th percentile: 1.8706583976745605ms<br/>Max time: 2.1457672119140625ms | Total: 1348.6347198486328ms<br/>Min time: 1.1456012725830078ms<br/>Mean time: 1.3486347198486328ms<br/>Median time: 1.2552738189697266ms<br/>95th percentile: 1.7621636390686035ms<br/>99th percentile: 1.8320250511169434ms<br/>Max time: 2.5415420532226562ms |
| e2e_route_ch | Total: 3234.7960472106934ms<br/>Min time: 1.3985633850097656ms<br/>Mean time: 3.2347960472106934ms<br/>Median time: 3.265500068664551ms<br/>95th percentile: 4.28394079208374ms<br/>99th percentile: 4.755227565765381ms<br/>Max time: 5.116701126098633ms | Total: 3082.5393199920654ms<br/>Min time: 1.3756752014160156ms<br/>Mean time: 3.0825393199920654ms<br/>Median time: 3.091573715209961ms<br/>95th percentile: 4.070770740509033ms<br/>99th percentile: 4.457108974456787ms<br/>Max time: 4.968166351318359ms |
| e2e_route_mld | Total: 3919.524908065796ms<br/>Min time: 1.363992691040039ms<br/>Mean time: 3.919524908065796ms<br/>Median time: 3.955721855163574ms<br/>95th percentile: 5.4238200187683105ms<br/>99th percentile: 5.954761505126953ms<br/>Max time: 6.522178649902344ms | Total: 3673.9938259124756ms<br/>Min time: 1.33514404296875ms<br/>Mean time: 3.6739938259124756ms<br/>Median time: 3.7049055099487305ms<br/>95th percentile: 5.030214786529541ms<br/>99th percentile: 5.600459575653076ms<br/>Max time: 6.049871444702148ms |
| e2e_table_ch | Total: 16721.204042434692ms<br/>Min time: 2.168893814086914ms<br/>Mean time: 16.721204042434692ms<br/>Median time: 16.142725944519043ms<br/>95th percentile: 30.30238151550293ms<br/>99th percentile: 32.10890769958496ms<br/>Max time: 33.35070610046387ms | Total: 16943.7358379364ms<br/>Min time: 2.1049976348876953ms<br/>Mean time: 16.9437358379364ms<br/>Median time: 16.260862350463867ms<br/>95th percentile: 31.094551086425778ms<br/>99th percentile: 32.82963514328003ms<br/>Max time: 34.83891487121582ms |
| e2e_table_mld | Total: 66201.20787620544ms<br/>Min time: 4.337787628173828ms<br/>Mean time: 66.20120787620544ms<br/>Median time: 62.483787536621094ms<br/>95th percentile: 127.60276794433594ms<br/>99th percentile: 135.3784418106079ms<br/>Max time: 147.04203605651855ms | Total: 64271.94929122925ms<br/>Min time: 3.998994827270508ms<br/>Mean time: 64.27194929122925ms<br/>Median time: 61.057209968566895ms<br/>95th percentile: 123.31310510635376ms<br/>99th percentile: 131.16525173187256ms<br/>Max time: 143.17655563354492ms |
| e2e_trip_ch | Total: 11609.169483184814ms<br/>Min time: 1.7392635345458984ms<br/>Mean time: 11.609169483184814ms<br/>Median time: 11.113524436950684ms<br/>95th percentile: 19.94059085845947ms<br/>99th percentile: 21.84530258178711ms<br/>Max time: 27.46438980102539ms | Total: 11530.039310455322ms<br/>Min time: 1.8410682678222656ms<br/>Mean time: 11.530039310455322ms<br/>Median time: 11.003255844116211ms<br/>95th percentile: 19.949376583099365ms<br/>99th percentile: 21.657054424285885ms<br/>Max time: 22.974014282226562ms |
| e2e_trip_mld | Total: 19491.829872131348ms<br/>Min time: 2.0856857299804688ms<br/>Mean time: 19.491829872131348ms<br/>Median time: 19.161105155944824ms<br/>95th percentile: 31.11332654953003ms<br/>99th percentile: 34.43922996520996ms<br/>Max time: 43.202877044677734ms | Total: 17912.116050720215ms<br/>Min time: 1.561880111694336ms<br/>Mean time: 17.912116050720215ms<br/>Median time: 17.40849018096924ms<br/>95th percentile: 29.06209230422973ms<br/>99th percentile: 31.29840850830078ms<br/>Max time: 33.9810848236084ms |
| json-render | String: 6.68917ms<br/>Stringstream: 9.41ms<br/>Vector: 7.03913ms | String: 6.74498ms<br/>Stringstream: 9.44459ms<br/>Vector: 6.98597ms |
| match_ch | Default radius: <br/>4.49425ms/req at 82 coordinate<br/>0.0548079ms/coordinate<br/>Radius 5m: <br/>4.47996ms/req at 82 coordinate<br/>0.0546337ms/coordinate<br/>Radius 10m: <br/>15.3433ms/req at 82 coordinate<br/>0.187113ms/coordinate<br/>Radius 15m: <br/>37.4741ms/req at 82 coordinate<br/>0.457001ms/coordinate<br/>Radius 30m: <br/>318.84ms/req at 82 coordinate<br/>3.88829ms/coordinate | Default radius: <br/>4.4448ms/req at 82 coordinate<br/>0.0542049ms/coordinate<br/>Radius 5m: <br/>4.40868ms/req at 82 coordinate<br/>0.0537644ms/coordinate<br/>Radius 10m: <br/>15.065ms/req at 82 coordinate<br/>0.183719ms/coordinate<br/>Radius 15m: <br/>37.1062ms/req at 82 coordinate<br/>0.452515ms/coordinate<br/>Radius 30m: <br/>316.308ms/req at 82 coordinate<br/>3.85741ms/coordinate |
| match_mld | Default radius: <br/>3.26618ms/req at 82 coordinate<br/>0.0398315ms/coordinate<br/>Radius 5m: <br/>2.83538ms/req at 82 coordinate<br/>0.0345778ms/coordinate<br/>Radius 10m: <br/>10.1179ms/req at 82 coordinate<br/>0.123389ms/coordinate<br/>Radius 15m: <br/>25.932ms/req at 82 coordinate<br/>0.316244ms/coordinate<br/>Radius 30m: <br/>303.134ms/req at 82 coordinate<br/>3.69676ms/coordinate | Default radius: <br/>2.78497ms/req at 82 coordinate<br/>0.033963ms/coordinate<br/>Radius 5m: <br/>2.77967ms/req at 82 coordinate<br/>0.0338984ms/coordinate<br/>Radius 10m: <br/>10.1795ms/req at 82 coordinate<br/>0.12414ms/coordinate<br/>Radius 15m: <br/>26.1554ms/req at 82 coordinate<br/>0.318968ms/coordinate<br/>Radius 30m: <br/>305.163ms/req at 82 coordinate<br/>3.72149ms/coordinate |
| osrm_contract | Time: 94.99s Peak RAM: 185.61MB | Time: 95.45s Peak RAM: 185.61MB |
| osrm_customize | Time: 1.34s Peak RAM: 115.12MB | Time: 1.32s Peak RAM: 115.13MB |
| osrm_extract | Time: 12.14s Peak RAM: 415.78MB | Time: 12.53s Peak RAM: 413.76MB |
| osrm_partition | Time: 2.21s Peak RAM: 148.44MB | Time: 2.30s Peak RAM: 148.37MB |
| packedvector | random write:<br/>std::vector 10189.3 ms<br/>util::packed_vector 82231.6 ms<br/>slowdown: 8.07042<br/>random read:<br/>std::vector 8688.59 ms<br/>util::packed_vector 33506.4 ms<br/>slowdown: 3.85637 | random write:<br/>std::vector 10333.1 ms<br/>util::packed_vector 81684.3 ms<br/>slowdown: 7.90508<br/>random read:<br/>std::vector 8628.23 ms<br/>util::packed_vector 30456.2 ms<br/>slowdown: 3.52984 |
| random_match_ch | 1000 matches, default radius<br/>total: 6846.82ms<br/>avg: 6.85ms<br/>min: 0.00ms<br/>max: 455.21ms<br/>p99: 105.79ms<br/><br/>1000 matches, radius=10<br/>total: 33970.23ms<br/>avg: 33.97ms<br/>min: 0.00ms<br/>max: 1779.31ms<br/>p99: 1762.44ms<br/><br/>1000 matches, radius=20<br/>total: 67049.58ms<br/>avg: 67.05ms<br/>min: 0.00ms<br/>max: 8695.85ms<br/>p99: 1157.31ms | 1000 matches, default radius<br/>total: 6745.32ms<br/>avg: 6.75ms<br/>min: 0.00ms<br/>max: 463.92ms<br/>p99: 102.77ms<br/><br/>1000 matches, radius=10<br/>total: 33727.32ms<br/>avg: 33.73ms<br/>min: 0.00ms<br/>max: 1835.39ms<br/>p99: 1804.49ms<br/><br/>1000 matches, radius=20<br/>total: 66074.55ms<br/>avg: 66.07ms<br/>min: 0.00ms<br/>max: 8997.14ms<br/>p99: 1155.91ms |
| random_match_mld | 1000 matches, default radius<br/>total: 5139.02ms<br/>avg: 5.14ms<br/>min: 0.00ms<br/>max: 387.30ms<br/>p99: 69.35ms<br/><br/>1000 matches, radius=10<br/>total: 26742.50ms<br/>avg: 26.74ms<br/>min: 0.00ms<br/>max: 1570.90ms<br/>p99: 1542.03ms<br/><br/>1000 matches, radius=20<br/>total: 52552.08ms<br/>avg: 52.55ms<br/>min: 0.00ms<br/>max: 6995.52ms<br/>p99: 797.74ms | 1000 matches, default radius<br/>total: 5152.87ms<br/>avg: 5.15ms<br/>min: 0.00ms<br/>max: 388.20ms<br/>p99: 70.24ms<br/><br/>1000 matches, radius=10<br/>total: 26571.44ms<br/>avg: 26.57ms<br/>min: 0.00ms<br/>max: 1548.20ms<br/>p99: 1532.11ms<br/><br/>1000 matches, radius=20<br/>total: 52620.93ms<br/>avg: 52.62ms<br/>min: 0.00ms<br/>max: 6986.87ms<br/>p99: 812.72ms |
| random_nearest_ch | 10000 nearest, number_of_results=1<br/>total: 484.69ms<br/>avg: 0.05ms<br/>min: 0.01ms<br/>max: 0.28ms<br/>p99: 0.12ms<br/><br/>10000 nearest, number_of_results=5<br/>total: 625.00ms<br/>avg: 0.06ms<br/>min: 0.02ms<br/>max: 0.28ms<br/>p99: 0.13ms<br/><br/>10000 nearest, number_of_results=10<br/>total: 774.95ms<br/>avg: 0.08ms<br/>min: 0.03ms<br/>max: 0.19ms<br/>p99: 0.15ms | 10000 nearest, number_of_results=1<br/>total: 478.18ms<br/>avg: 0.05ms<br/>min: 0.01ms<br/>max: 0.26ms<br/>p99: 0.12ms<br/><br/>10000 nearest, number_of_results=5<br/>total: 618.19ms<br/>avg: 0.06ms<br/>min: 0.02ms<br/>max: 0.21ms<br/>p99: 0.13ms<br/><br/>10000 nearest, number_of_results=10<br/>total: 772.21ms<br/>avg: 0.08ms<br/>min: 0.04ms<br/>max: 0.19ms<br/>p99: 0.14ms |
| random_nearest_mld | 10000 nearest, number_of_results=1<br/>total: 481.68ms<br/>avg: 0.05ms<br/>min: 0.01ms<br/>max: 0.28ms<br/>p99: 0.12ms<br/><br/>10000 nearest, number_of_results=5<br/>total: 615.02ms<br/>avg: 0.06ms<br/>min: 0.02ms<br/>max: 0.23ms<br/>p99: 0.13ms<br/><br/>10000 nearest, number_of_results=10<br/>total: 771.15ms<br/>avg: 0.08ms<br/>min: 0.04ms<br/>max: 0.19ms<br/>p99: 0.14ms | 10000 nearest, number_of_results=1<br/>total: 475.60ms<br/>avg: 0.05ms<br/>min: 0.01ms<br/>max: 0.28ms<br/>p99: 0.12ms<br/><br/>10000 nearest, number_of_results=5<br/>total: 616.82ms<br/>avg: 0.06ms<br/>min: 0.02ms<br/>max: 0.21ms<br/>p99: 0.13ms<br/><br/>10000 nearest, number_of_results=10<br/>total: 766.54ms<br/>avg: 0.08ms<br/>min: 0.04ms<br/>max: 0.19ms<br/>p99: 0.14ms |
| random_route_ch | 10000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>total: 21933.86ms<br/>avg: 2.19ms<br/>min: 0.16ms<br/>max: 3.92ms<br/>p99: 3.31ms<br/><br/>10000 routes, 2 coordinates, no alternatives, overview=full, steps=true<br/>total: 10126.78ms<br/>avg: 1.01ms<br/>min: 0.07ms<br/>max: 2.86ms<br/>p99: 1.69ms<br/><br/>10000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>total: 19117.57ms<br/>avg: 1.91ms<br/>min: 0.09ms<br/>max: 5.72ms<br/>p99: 4.19ms<br/><br/>10000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>total: 10454.40ms<br/>avg: 1.05ms<br/>min: 0.12ms<br/>max: 1.96ms<br/>p99: 1.51ms<br/><br/>10000 routes, 2 coordinates, no alternatives, overview=false, steps=false<br/>total: 4443.96ms<br/>avg: 0.44ms<br/>min: 0.05ms<br/>max: 0.91ms<br/>p99: 0.70ms<br/><br/>10000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>total: 9966.79ms<br/>avg: 1.00ms<br/>min: 0.06ms<br/>max: 4.36ms<br/>p99: 2.21ms<br/><br/>10000 routes, 3 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>total: 692.25ms<br/>avg: 0.07ms<br/>min: 0.01ms<br/>max: 1.65ms<br/>p99: 0.70ms<br/><br/>10000 routes, 2 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>total: 705.54ms<br/>avg: 0.07ms<br/>min: 0.01ms<br/>max: 0.55ms<br/>p99: 0.42ms<br/><br/>10000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false, radius=750<br/>total: 932.00ms<br/>avg: 0.09ms<br/>min: 0.01ms<br/>max: 1.86ms<br/>p99: 1.14ms | 10000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>total: 22926.71ms<br/>avg: 2.29ms<br/>min: 0.17ms<br/>max: 5.95ms<br/>p99: 3.56ms<br/><br/>10000 routes, 2 coordinates, no alternatives, overview=full, steps=true<br/>total: 10577.34ms<br/>avg: 1.06ms<br/>min: 0.07ms<br/>max: 2.46ms<br/>p99: 1.79ms<br/><br/>10000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>total: 20319.55ms<br/>avg: 2.03ms<br/>min: 0.09ms<br/>max: 5.69ms<br/>p99: 4.73ms<br/><br/>10000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>total: 10937.74ms<br/>avg: 1.09ms<br/>min: 0.12ms<br/>max: 3.25ms<br/>p99: 1.63ms<br/><br/>10000 routes, 2 coordinates, no alternatives, overview=false, steps=false<br/>total: 4411.31ms<br/>avg: 0.44ms<br/>min: 0.05ms<br/>max: 0.83ms<br/>p99: 0.68ms<br/><br/>10000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>total: 11145.01ms<br/>avg: 1.11ms<br/>min: 0.06ms<br/>max: 5.98ms<br/>p99: 2.99ms<br/><br/>10000 routes, 3 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>total: 706.72ms<br/>avg: 0.07ms<br/>min: 0.01ms<br/>max: 1.56ms<br/>p99: 0.77ms<br/><br/>10000 routes, 2 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>total: 731.15ms<br/>avg: 0.07ms<br/>min: 0.01ms<br/>max: 0.61ms<br/>p99: 0.48ms<br/><br/>10000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false, radius=750<br/>total: 993.51ms<br/>avg: 0.10ms<br/>min: 0.01ms<br/>max: 2.08ms<br/>p99: 1.35ms |
| random_route_mld | 10000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>total: 42907.55ms<br/>avg: 4.29ms<br/>min: 0.14ms<br/>max: 10.77ms<br/>p99: 7.40ms<br/><br/>10000 routes, 2 coordinates, no alternatives, overview=full, steps=true<br/>total: 15505.67ms<br/>avg: 1.55ms<br/>min: 0.09ms<br/>max: 3.31ms<br/>p99: 2.82ms<br/><br/>10000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>total: 42901.79ms<br/>avg: 4.29ms<br/>min: 0.07ms<br/>max: 11.68ms<br/>p99: 8.85ms<br/><br/>10000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>total: 30374.93ms<br/>avg: 3.04ms<br/>min: 0.12ms<br/>max: 9.45ms<br/>p99: 5.41ms<br/><br/>10000 routes, 2 coordinates, no alternatives, overview=false, steps=false<br/>total: 8916.93ms<br/>avg: 0.89ms<br/>min: 0.05ms<br/>max: 1.88ms<br/>p99: 1.58ms<br/><br/>10000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>total: 32699.92ms<br/>avg: 3.27ms<br/>min: 0.05ms<br/>max: 10.09ms<br/>p99: 6.66ms<br/><br/>10000 routes, 3 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>total: 860.36ms<br/>avg: 0.09ms<br/>min: 0.01ms<br/>max: 4.36ms<br/>p99: 1.56ms<br/><br/>10000 routes, 2 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>total: 943.63ms<br/>avg: 0.09ms<br/>min: 0.01ms<br/>max: 1.27ms<br/>p99: 1.02ms<br/><br/>10000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false, radius=750<br/>total: 1725.02ms<br/>avg: 0.17ms<br/>min: 0.01ms<br/>max: 4.97ms<br/>p99: 3.48ms | 10000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>total: 42957.60ms<br/>avg: 4.30ms<br/>min: 0.13ms<br/>max: 10.40ms<br/>p99: 7.43ms<br/><br/>10000 routes, 2 coordinates, no alternatives, overview=full, steps=true<br/>total: 15316.49ms<br/>avg: 1.53ms<br/>min: 0.09ms<br/>max: 3.35ms<br/>p99: 2.78ms<br/><br/>10000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>total: 43816.40ms<br/>avg: 4.38ms<br/>min: 0.09ms<br/>max: 11.46ms<br/>p99: 9.26ms<br/><br/>10000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>total: 31607.55ms<br/>avg: 3.16ms<br/>min: 0.11ms<br/>max: 9.66ms<br/>p99: 5.66ms<br/><br/>10000 routes, 2 coordinates, no alternatives, overview=false, steps=false<br/>total: 10364.64ms<br/>avg: 1.04ms<br/>min: 0.05ms<br/>max: 2.34ms<br/>p99: 1.95ms<br/><br/>10000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>total: 35517.73ms<br/>avg: 3.55ms<br/>min: 0.05ms<br/>max: 10.95ms<br/>p99: 7.37ms<br/><br/>10000 routes, 3 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>total: 895.90ms<br/>avg: 0.09ms<br/>min: 0.01ms<br/>max: 4.66ms<br/>p99: 1.70ms<br/><br/>10000 routes, 2 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>total: 1037.21ms<br/>avg: 0.10ms<br/>min: 0.01ms<br/>max: 1.76ms<br/>p99: 1.25ms<br/><br/>10000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false, radius=750<br/>total: 1871.63ms<br/>avg: 0.19ms<br/>min: 0.01ms<br/>max: 5.40ms<br/>p99: 3.84ms |
| random_table_ch | 250 tables, 3 coordinates<br/>total: 200.22ms<br/>avg: 0.80ms<br/>min: 0.51ms<br/>max: 2.15ms<br/>p99: 1.53ms<br/><br/>250 tables, 25 coordinates<br/>total: 1481.25ms<br/>avg: 5.93ms<br/>min: 5.40ms<br/>max: 6.41ms<br/>p99: 6.38ms<br/><br/>250 tables, 50 coordinates<br/>total: 3019.07ms<br/>avg: 12.08ms<br/>min: 11.01ms<br/>max: 17.97ms<br/>p99: 15.17ms<br/><br/>250 tables, 100 coordinates<br/>total: 6467.04ms<br/>avg: 25.87ms<br/>min: 24.42ms<br/>max: 27.94ms<br/>p99: 27.39ms | 250 tables, 3 coordinates<br/>total: 189.01ms<br/>avg: 0.76ms<br/>min: 0.57ms<br/>max: 1.91ms<br/>p99: 1.02ms<br/><br/>250 tables, 25 coordinates<br/>total: 1568.02ms<br/>avg: 6.27ms<br/>min: 5.76ms<br/>max: 6.71ms<br/>p99: 6.68ms<br/><br/>250 tables, 50 coordinates<br/>total: 3166.46ms<br/>avg: 12.67ms<br/>min: 11.81ms<br/>max: 13.39ms<br/>p99: 13.29ms<br/><br/>250 tables, 100 coordinates<br/>total: 6852.31ms<br/>avg: 27.41ms<br/>min: 26.20ms<br/>max: 30.31ms<br/>p99: 28.40ms |
| random_table_mld | 250 tables, 3 coordinates<br/>total: 764.28ms<br/>avg: 3.06ms<br/>min: 2.37ms<br/>max: 4.31ms<br/>p99: 4.03ms<br/><br/>250 tables, 25 coordinates<br/>total: 7053.81ms<br/>avg: 28.22ms<br/>min: 24.96ms<br/>max: 32.99ms<br/>p99: 31.99ms<br/><br/>250 tables, 50 coordinates<br/>total: 14998.69ms<br/>avg: 59.99ms<br/>min: 55.40ms<br/>max: 78.96ms<br/>p99: 65.22ms<br/><br/>250 tables, 100 coordinates<br/>total: 32802.42ms<br/>avg: 131.21ms<br/>min: 123.62ms<br/>max: 147.78ms<br/>p99: 141.38ms | 250 tables, 3 coordinates<br/>total: 741.58ms<br/>avg: 2.97ms<br/>min: 2.38ms<br/>max: 4.19ms<br/>p99: 3.94ms<br/><br/>250 tables, 25 coordinates<br/>total: 6927.92ms<br/>avg: 27.71ms<br/>min: 24.76ms<br/>max: 37.62ms<br/>p99: 33.10ms<br/><br/>250 tables, 50 coordinates<br/>total: 14719.00ms<br/>avg: 58.88ms<br/>min: 54.63ms<br/>max: 64.13ms<br/>p99: 62.30ms<br/><br/>250 tables, 100 coordinates<br/>total: 31983.98ms<br/>avg: 127.94ms<br/>min: 122.42ms<br/>max: 147.42ms<br/>p99: 135.84ms |
…